pyproject.toml
[tool.poetry]
name = "dad-joke-bot"
version = "0.1.0"
description = ""
authors = ["Sam Atkins <samatkins@outlook.com>"]

[tool.poetry.dependencies]
python = "3.6.6"
python-dotenv = "^0.9.1"
zappa = "^0.47.0"
requests = "^2.20"
python-twitter = "^3.5"


[tool.poetry.dev-dependencies]
pytest = "^3.0"
pytest-cov = "^2.5"
yapf = "^0.23.0"
flake8 = "^3.5"
